+++ mingling/src/res.rs
@@ -1,5 +1,8 @@
 mod exit_code;
 pub use exit_code::*;
 
+mod dirs;
+pub use dirs::*;
+
 #[allow(unused_imports)]
 pub use mingling_core::core_res::*;