+++ mingling/src/res/dirs.rs
@@ -1,0 +1,11 @@
+mod current_dir;
+pub use current_dir::*;
+
+mod current_exe;
+pub use current_exe::*;
+
+mod temp_dir;
+pub use temp_dir::*;
+
+mod home_dir;
+pub use home_dir::*;